Fix: Make !remind Compatible With Multi-Word Roles (#35)
* Added Regex Checks

The new Regex checks check for words within two ` symbols and everything after. These are then placed into the paramArray as the role selected and the message to send

* fix: compile errors & console.log() cleanup

The first warning can be prevented if we just add the `?` to check if the value is null before indexing

The second issue is just a matter of forcing a toString() conversion. No harm is done since we know it won't be null (based off of the first check)

* fix: accepting roles with slashes

This should allow commands like !remind `he/him` hello!

* fix: regex matching & message styles

Co-authored-by: Vlad <vladjdk@gmail.com>
Co-authored-by: aMahanna <anthony.mahanna@gmail.com>
Co-authored-by: Ryan Fleck <ryan.fleck@protonmail.com>